Extract admission tests into admission/tests.rs
